Add files via upload
Cleaned up the code
